// models/sales_envelope.model.ts
export type SalesEnvelopeState = {
  data: {
    sals_enve_id: string;
    sals_enve_date: string;
    sals_enve_fullname: string;
    sals_enve_company_name: string;
    sals_enve_tel: string;
    sals_enve_email: string;
    sals_enve_doc_type: string;
    sals_enve_printing_type: string;
    sals_enve_amount: string;
    sals_enve_quotation_request: string;
    sals_enve_finished_size: string;
    sals_enve_page: string;
    sals_enve_paper: string;
    sals_enve_printing: string;
    sals_enve_coating: string;
    sals_enve_printing_volume: string;
  };
};